Move Kokkos_Serial files to Serial directory
